requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
temporalio==1.17.0
redis==7.0.0
python-dotenv==1.1.0
weaviate-client==4.17.0
pytest==8.3.4
pytest-asyncio==0.24.0
langchain==1.0.8
langchain-community==0.4.1
langchain-core==1.0.7
langchain-google-vertexai==3.0.3
langchain-google-genai
langchain-ollama
langchain-openai==1.0.3
httpx==0.28.0
pydantic==2.12.0
markdown==3.9
python-json-logger==2.0.7
pymongo==4.10.1
rank-bm25==0.2.2